requirements.txt
pyautogui>=0.9.54
rapidocr-onnxruntime>=1.2.3,<1.3
opencv-python==4.10.0.84; python_version < "3.9"
opencv-python>=4.9.0; python_version >= "3.9"
numpy==1.24.4; python_version < "3.9"
numpy>=1.26.0; python_version >= "3.9"
Pillow>=10.0.0
pyobjc-core==9.2; python_version < "3.9"
pyobjc-framework-Cocoa==9.2; python_version < "3.9"
pyobjc-framework-Quartz==9.2; python_version < "3.9"
pyobjc-framework-Quartz>=11.0; python_version >= "3.9"
pyperclip>=1.8.2
tomli>=2.0.1; python_version < "3.11"
certifi>=2024.8.30
